fix(agents): stop dumping raw CLI JSONL as autofix error messages
The autofixer's analysis/fix phases already summarize failed AI CLI
runs into a short, human-readable message (autofixer.rs), but three
other call sites still stored the raw stream-json/stdout dump as
error_message:

- ClaudeCliProvider::run/continue_conversation (ai_cli/claude.rs)
- OpenCodeCliProvider::run/continue_conversation (ai_cli/opencode.rs)
- CodexCliProvider::run stderr (ai_cli/codex.rs)
- AgentExecutor's generic workflow-run path (services/executor.rs)

These are reachable from the general-purpose AI agents/workflow
executor (AgentExecutor::with_ai_provider), not just the error-tracking
autofix button, so the same megabytes-of-JSON-in-an-Alert bug could
surface there too.

Extract the existing summarize_cli_failure/scrub_and_bound helpers out
of autofixer.rs into ai_cli (the lower-level module all providers and
the executor already depend on) and reuse them at all four remaining
sites instead of duplicating or re-inventing the logic.

Also harden the frontend as defense-in-depth, since error_message is
still rendered without any bound: AutofixButton's subtitle now clamps
to 2 lines/160 chars instead of dumping the raw string inline, and
AutopilotRunDetail's error Alert caps at 1200 chars with a 'View full'
dialog, matching the existing prompt_text truncation pattern in the
same file.
