+++ crates/temps-agents/src/ai_cli/opencode.rs
@@ -186,7 +186,7 @@ impl AiCliProvider for OpenCodeCliProvider {
             return Err(AgentError::AiCliFailed {
                 provider: self.name().to_string(),
                 exit_code,
-                stderr: error_output,
+                stderr: crate::ai_cli::summarize_cli_failure(self.name(), &error_output),
             });
         }
 
@@ -297,7 +297,7 @@ impl AiCliProvider for OpenCodeCliProvider {
             return Err(AgentError::AiCliFailed {
                 provider: self.name().to_string(),
                 exit_code,
-                stderr: error_output,
+                stderr: crate::ai_cli::summarize_cli_failure(self.name(), &error_output),
             });
         }
